style and performance fixes

+++ day03p1.jl
@@ -8,8 +8,8 @@ ans = 0
 open(filename, "r") do f
     for line in eachline(f)
         for m in eachmatch(exp, line)
-            n1 = parse(Int64, m.captures[1])
-            n2 = parse(Int64, m.captures[2])
+            n1 = parse(Int, m.captures[1])
+            n2 = parse(Int, m.captures[2])
             global ans += n1 * n2
         end
     end
